Split AdamRMS integration into separate permission
