+++ lib/auth/permissions.ts
@@ -7,9 +7,6 @@ import { z } from "zod";
  * * PUBLIC - open to the world with no authentication
  * * SuperUser - can do anything (don't use this unless you know what you're doing)
  */
-// TODO: This is duplicated between here and the DB. In theory we could just use `string` as the type, but that
-//  loses auto-complete. We could also auto-generate it from the DB, but my preference would be to remove the
-//  DB permissions table entirely and have the codebase be the source of truth.
 export const PermissionEnum = z.enum([
   "PUBLIC",
   "MEMBER",
@@ -22,10 +19,9 @@ export const PermissionEnum = z.enum([
   "Calendar.Meeting.Creator",
   "Calendar.Social.Admin",
   "Calendar.Social.Creator",
+  "CalendarIntegration.Admin",
   "ManageMembers.Members.List",
   "ManageMembers.Members.Admin",
   "ManageMembers.Admin",
 ]);
 export type Permission = z.infer<typeof PermissionEnum>;
-
-export * from "./errors";
